requirements.txt
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
comtypes==1.2.0
google-ai-generativelanguage==0.3.3
google-api-core==2.12.0
google-auth==2.23.3
google-generativeai==0.2.2
googleapis-common-protos==1.61.0
grpcio==1.59.0
grpcio-status==1.59.0
gTTS==2.4.0
idna==3.4
proto-plus==1.22.3
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
pygame==2.5.2
pypiwin32==223
pyttsx3==2.90
pywin32==306
requests==2.31.0
rsa==4.9
SpeechRecognition==3.10.0
tqdm==4.66.1
urllib3==2.0.7
